test: migrate integration tests
