+++ test/integration.test.ts
@@ -6,7 +6,6 @@ import {
   existsFile,
   assertFilesContent,
   assertContainFiles,
-  deleteFile,
 } from './testing-utils'
 import * as debug from './utils/debug'
 
@@ -440,158 +439,6 @@ const testCases: {
       expect(indexLine.indexOf('.')).toEqual(exportsIndex)
       expect(indexLine.indexOf('dist/index.js')).toEqual(fileIndex)
       expect(getOutputSizeColumnIndex(indexLine)).toEqual(sizeIndex)
-    },
-  },
-  {
-    name: 'prepare-js',
-    args: ['--prepare'],
-    async before(dir) {
-      await fsp.writeFile(join(dir, './package.json'), '{ "type": "commonjs" }')
-    },
-    async expected(dir, { stdout }) {
-      assertContainFiles(dir, ['package.json'])
-      const pkgJson = JSON.parse(
-        await fsp.readFile(join(dir, './package.json'), 'utf-8'),
-      )
-      expect(pkgJson.main).toBe('./dist/index.js')
-      expect(pkgJson.module).toBe('./dist/index.mjs')
-      expect(pkgJson.types).toBeFalsy()
-      expect(pkgJson.files).toContain('dist')
-      expect(pkgJson.bin).toBe('./dist/bin/index.js')
-      expect(pkgJson.exports).toEqual({
-        './foo': {
-          import: './dist/foo.mjs',
-          require: './dist/foo.js',
-        },
-        '.': {
-          import: './dist/index.mjs',
-          require: './dist/index.js',
-        },
-      })
-
-      /*
-        Discovered binaries entries:
-          .: bin.js
-        Discovered exports entries:
-          ./foo: foo.js
-          .    : index.js
-        ✓ Configured `exports` in package.json
-      */
-      expect(stdout).toContain('Discovered binaries entries:')
-      expect(stdout).toMatch(/.\s*: bin.js/)
-      expect(stdout).toContain('Discovered exports entries:')
-      expect(stdout).toMatch(/\.\/foo\s*: foo.js/)
-      expect(stdout).toMatch(/\.\s*: index.js/)
-      expect(stripANSIColor(stdout)).toContain(
-        '✓ Configured `exports` in package.json',
-      )
-    },
-  },
-  {
-    name: 'prepare-ts',
-    args: ['--prepare'],
-    async before(dir) {
-      await deleteFile(join(dir, './package.json'))
-      await deleteFile(join(dir, './tsconfig.json'))
-    },
-    async expected(dir, { stdout }) {
-      assertContainFiles(dir, ['package.json'])
-      const pkgJson = JSON.parse(
-        await fsp.readFile(join(dir, './package.json'), 'utf-8'),
-      )
-      expect(pkgJson.files).toContain('dist')
-      expect(pkgJson.bin).toEqual({
-        cli: './dist/bin/cli.js',
-        cmd: './dist/bin/cmd.js',
-      })
-      expect(pkgJson.type).toBe('module')
-      expect(pkgJson.main).toBe('./dist/es/index.js')
-      expect(pkgJson.module).toBe('./dist/es/index.js')
-      expect(pkgJson.exports).toEqual({
-        './foo': {
-          import: {
-            types: './dist/es/foo.d.ts',
-            default: './dist/es/foo.js',
-          },
-          require: {
-            types: './dist/cjs/foo.d.cts',
-            default: './dist/cjs/foo.cjs',
-          },
-        },
-        '.': {
-          'react-server': './dist/es/index-react-server.mjs',
-          import: {
-            types: './dist/es/index.d.ts',
-            default: './dist/es/index.js',
-          },
-          require: {
-            types: './dist/cjs/index.d.cts',
-            default: './dist/cjs/index.cjs',
-          },
-        },
-      })
-
-      /*
-        Discovered binaries entries:
-          ./cli: cli.ts
-          ./cmd: cmd.ts
-        Discovered exports entries:
-          ./foo: foo.ts
-          .    : index.react-server.ts
-          .    : index.ts
-        ✓ Configured `exports` in package.json
-      */
-      expect(stripANSIColor(stdout)).toContain(
-        'Detected using TypeScript but tsconfig.json is missing, created a tsconfig.json for you.',
-      )
-      expect(stdout).toContain('Discovered binaries entries:')
-      expect(stdout).toMatch(/\.\s*: index.ts/)
-      expect(stdout).toContain('Discovered exports entries:')
-      expect(stdout).toMatch(/\.\/foo\s*: foo.ts/)
-      expect(stdout).toMatch(/\.\s*: index.react-server.ts/)
-      expect(stripANSIColor(stdout)).toContain(
-        '✓ Configured `exports` in package.json',
-      )
-    },
-  },
-  {
-    name: 'prepare-ts-with-test-file',
-    args: ['--prepare'],
-    async before(dir) {
-      await deleteFile(join(dir, './package.json'))
-      await deleteFile(join(dir, './tsconfig.json'))
-    },
-    async expected(dir) {
-      assertContainFiles(dir, ['package.json'])
-      const pkgJson = JSON.parse(
-        await fsp.readFile(join(dir, './package.json'), 'utf-8'),
-      )
-      expect(pkgJson.files).toContain('dist')
-      expect(pkgJson.main).toBe('./dist/es/index.js')
-      expect(pkgJson.module).toBe('./dist/es/index.js')
-      expect(Object.keys(pkgJson.exports)).toEqual(['.'])
-      expect(Object.keys(pkgJson.exports['.'])).not.toContain('./test')
-    },
-  },
-  {
-    name: 'prepare-ts-with-pkg-json',
-    args: ['--prepare'],
-    async before(dir) {
-      await fsp.writeFile(
-        join(dir, './package.json'),
-        '{ "name": "prepare-ts-with-pkg-json" }',
-      )
-      await deleteFile(join(dir, './tsconfig.json'))
-    },
-    async expected(dir) {
-      assertContainFiles(dir, ['package.json'])
-      const pkgJson = JSON.parse(
-        await fsp.readFile(join(dir, './package.json'), 'utf-8'),
-      )
-      expect(pkgJson.files).toContain('dist')
-      expect(pkgJson.type).toBeUndefined()
-      expect(pkgJson.main).toBe('./dist/cjs/index.js')
-      expect(pkgJson.module).toBe('./dist/es/index.mjs')
     },
   },
   {
